stream from a named pipe
Create a FIFO (a named pipe) to write to and read from simultaneously
with ffmpeg, allowing a continuous stream instead of restarting every
queue pop. The audio filter `anullsrc` is used to generate a silent
track while the queue is empty. Queue requests are padded with 5 seconds
of silence as a transition. Thanks to Slugalisk for the idea and help
with ffmpeg :)

Signed-off-by: jbpratt <jbpratt78@gmail.com>